Web App/tapband_server/flask_backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-SocketIO
Flask-Migrate
paho-mqtt
python-dotenv
